pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioreads"
version = "0.1.0"
description = "NGS quality control and alignment for multiple sequencing platforms"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.81",
    "click>=8.0",
    "customtkinter>=5.2",
    "tkinterdnd2>=0.3",
]

[project.optional-dependencies]
align = [
    "mappy>=2.24; sys_platform != 'win32'",
    "pysam>=0.22; sys_platform != 'win32'",
]
dev = ["pytest>=7.0"]

[project.scripts]
bioreads = "bioreads.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
